Exception Handling of DELETE

+++ src/main/java/com/starbuy/ecommerce/services/ProductService.java
@@ -4,14 +4,15 @@ import com.starbuy.ecommerce.dto.ProductDTO;
 import com.starbuy.ecommerce.dto.ProductRequestDTO;
 import com.starbuy.ecommerce.models.Product;
 import com.starbuy.ecommerce.repositories.ProductRepository;
+import com.starbuy.ecommerce.services.exceptions.DatabaseException;
 import com.starbuy.ecommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.beans.PropertyEditorSupport;
-import java.util.Optional;
 
 @Service
 public class ProductService {
@@ -46,15 +47,28 @@ public class ProductService {
 
     @Transactional
     public ProductDTO updateProduct(Long id, ProductRequestDTO dto){
-        Product product = repository.getReferenceById(id);
-        copyDtoToEntity(dto, product);
-        product = repository.save(product);
-        return new ProductDTO(product);
+        try{
+            Product product = repository.getReferenceById(id);
+            copyDtoToEntity(dto, product);
+            product = repository.save(product);
+            return new ProductDTO(product);
+        }
+        catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Id Not Found");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void deleteById(Long id){
-        repository.deleteById(id);
+        if (!repository.existsById(id)){
+            throw new ResourceNotFoundException("Id Not Found");
+        }
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e){
+            throw new DatabaseException("Failure in Referential Integrity");
+        }
     }
 
     private void copyDtoToEntity(ProductRequestDTO dto, Product product) {
